backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1
openai==1.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
Pillow==10.2.0
pypdf==4.0.1
httpx==0.26.0
python-docx>=1.1.0
extract-msg>=0.41.0
# macOS Vision OCR (for document viewer with field highlighting)
pymupdf==1.24.0
pyobjc-framework-Vision==10.1
pyobjc-framework-Quartz==10.1
